ToolbarItem changed to Command, move settings to main app
