+++ Safari Extension/Info.plist
@@ -45,7 +45,7 @@
 		<key>SFSafariToolbarItem</key>
 		<dict>
 			<key>Action</key>
-			<string>Popover</string>
+			<string>Command</string>
 			<key>Identifier</key>
 			<string>Button</string>
 			<key>Image</key>
